Stage insight builder for move to test utilities
JIRA: RAIL-3492
